First working version of the plugin
This commits adds a settings page for this plugin under WooCommerce admin menu. This page displays a single button that when clicked you add a custom meta_value/meta_key index to wp_postmeta.

This is just the very basic functionality of the plugin: queries are run in the same request instead of a background process, no feedback is provided to the user, errors are not handled and the site is not put in maintenance mode while queries are executing.